requirements.txt
Faker==37.3.0
kafka-python==2.2.7
tzdata==2025.2
